pyproject.toml
[tool.poetry]
name = "book-shop-api"
version = "0.1.0"
description = ""
authors = ["nuriddin <sobitovnuriddin7655@gmail.com>"]
readme = "README.md"
packages = [{include = "book_shop_api"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.3"
django-mptt = "^0.14.0"
django-rest-framework = "^0.1.0"
pdfreader = "^0.1.12"
pypdf2 = "^3.0.1"
django-filter = "^23.2"
django-rest-swagger = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
